refactor(commons): restructured packages
