--- commons/src/main/kotlin/piperkt/common/DomainEvent.kt
@@ -1,3 +1,0 @@
-package piperkt.common
-
-interface DomainEvent : ValueObject
