--- commons/src/main/kotlin/piperkt/common/UseCase.kt
@@ -1,15 +1,0 @@
-package piperkt.common
-
-interface QueryUseCase<in Q, out R> {
-
-    fun execute(query: Q): R
-
-    operator fun invoke(query: Q): Result<R> = runCatching { execute(query) }
-}
-
-interface CommandUseCase<in C> {
-
-    fun execute(command: C)
-
-    operator fun invoke(command: C): Result<Unit> = runCatching { execute(command) }
-}
