+++ commons/src/main/kotlin/piperkt/common/events/ChannelEvent.kt
@@ -1,8 +1,5 @@
 package piperkt.common.events
 
-import piperkt.common.DomainEvent
-import piperkt.common.EventPublisher
-
 sealed interface ChannelEvent : DomainEvent {
     data class ChannelCreatedEvent(val serverId: String, val channelId: String) : ChannelEvent
