+++ commons/src/main/kotlin/piperkt/common/events/DomainEvent.kt
@@ -1,0 +1,5 @@
+package piperkt.common.events
+
+import piperkt.common.ddd.ValueObject
+
+interface DomainEvent : ValueObject
