+++ commons/src/main/kotlin/piperkt/common/events/EventListener.kt
@@ -1,4 +1,4 @@
-package piperkt.common
+package piperkt.common.events
 
 interface EventListener<E : DomainEvent> {
     fun handle(event: E)
