+++ commons/src/main/kotlin/piperkt/common/events/EventPublisher.kt
@@ -1,4 +1,4 @@
-package piperkt.common
+package piperkt.common.events
 
 interface EventPublisher<E : DomainEvent> {
     fun publish(event: E)
