+++ commons/src/main/kotlin/piperkt/common/id/FriendshipAggregateId.kt
@@ -1,5 +1,5 @@
 package piperkt.common.id
 
-import piperkt.common.UUIDEntityId
+import piperkt.common.ddd.UUIDEntityId
 
 class FriendshipAggregateId(value: String = newId()) : UUIDEntityId(value)
