+++ commons/src/main/kotlin/piperkt/common/id/FriendshipId.kt
@@ -1,5 +1,5 @@
 package piperkt.common.id
 
-import piperkt.common.UUIDEntityId
+import piperkt.common.ddd.UUIDEntityId
 
 class FriendshipId(value: String = newId()) : UUIDEntityId(value)
