+++ commons/src/main/kotlin/piperkt/common/id/MessageId.kt
@@ -1,5 +1,5 @@
 package piperkt.common.id
 
-import piperkt.common.UUIDEntityId
+import piperkt.common.ddd.UUIDEntityId
 
 class MessageId(value: String = newId()) : UUIDEntityId(value)
