+++ friendships-service/src/main/kotlin/piperkt/services/friendships/application/FriendshipRepository.kt
@@ -1,6 +1,6 @@
 package piperkt.services.friendships.application
 
-import piperkt.common.Repository
+import piperkt.common.ddd.Repository
 import piperkt.common.id.FriendshipId
 import piperkt.services.friendships.domain.Friendship
