+++ friendships-service/src/main/kotlin/piperkt/services/friendships/domain/FriendshipRequest.kt
@@ -1,6 +1,6 @@
 package piperkt.services.friendships.domain
 
-import piperkt.common.AggregateRoot
+import piperkt.common.ddd.AggregateRoot
 import piperkt.common.id.FriendshipRequestId
 
 class FriendshipRequest(
