+++ friendships-service/src/main/kotlin/piperkt/services/friendships/domain/Message.kt
@@ -1,7 +1,7 @@
 package piperkt.services.friendships.domain
 
 import java.time.Instant
-import piperkt.common.Entity
+import piperkt.common.ddd.Entity
 import piperkt.common.id.MessageId
 
 class Message(
