+++ multimedia-service/src/main/kotlin/piperkt/services/multimedia/application/direct/DirectEventsListener.kt
@@ -1,6 +1,6 @@
 package piperkt.services.multimedia.application.direct
 
-import piperkt.common.EventListener
+import piperkt.common.events.EventListener
 import piperkt.common.events.FriendshipEvent
 import piperkt.services.multimedia.application.session.SessionService
 import piperkt.services.multimedia.application.session.SessionService.Command.CreateSession
