+++ multimedia-service/src/main/kotlin/piperkt/services/multimedia/domain/direct/Direct.kt
@@ -1,7 +1,7 @@
 package piperkt.services.multimedia.domain.direct
 
-import piperkt.common.AggregateRoot
-import piperkt.common.UUIDEntityId
+import piperkt.common.ddd.AggregateRoot
+import piperkt.common.ddd.UUIDEntityId
 import piperkt.services.multimedia.domain.session.SessionId
 import piperkt.services.multimedia.domain.user.Username
 
@@ -10,5 +10,5 @@ class DirectId(value: String = newId()) : UUIDEntityId(value)
 class Direct(
     override val id: DirectId = DirectId(),
     val users: Set<Username>,
-    val sessionId: SessionId
+    val sessionId: SessionId,
 ) : AggregateRoot<DirectId>(id)
