+++ multimedia-service/src/main/kotlin/piperkt/services/multimedia/domain/server/ServerRepository.kt
@@ -1,5 +1,5 @@
 package piperkt.services.multimedia.domain.server
 
-import piperkt.common.Repository
+import piperkt.common.ddd.Repository
 
 interface ServerRepository : Repository<ServerId, Server>
