+++ multimedia-service/src/main/kotlin/piperkt/services/multimedia/domain/session/SessionRepository.kt
@@ -1,5 +1,5 @@
 package piperkt.services.multimedia.domain.session
 
-import piperkt.common.Repository
+import piperkt.common.ddd.Repository
 
 interface SessionRepository : Repository<SessionId, Session>
