+++ multimedia-service/src/main/kotlin/piperkt/services/multimedia/domain/user/User.kt
@@ -1,7 +1,7 @@
 package piperkt.services.multimedia.domain.user
 
-import piperkt.common.AggregateRoot
-import piperkt.common.EntityId
+import piperkt.common.ddd.AggregateRoot
+import piperkt.common.ddd.EntityId
 
 class Username(value: String) : EntityId<String>(value)
