+++ multimedia-service/src/test/kotlin/mocks/publishers/MockedSessionEventPublisher.kt
@@ -1,7 +1,7 @@
 package mocks.publishers
 
-import piperkt.common.MockedEventPublisher
 import piperkt.common.events.SessionEvent
 import piperkt.common.events.SessionEventPublisher
+import piperkt.common.mocks.MockedEventPublisher
 
 class MockedSessionEventPublisher : SessionEventPublisher, MockedEventPublisher<SessionEvent>()
