+++ servers-service/src/main/kotlin/piperkt/services/servers/domain/factory/ChannelFactory.kt
@@ -1,6 +1,6 @@
 package piperkt.services.servers.domain.factory
 
-import piperkt.common.Factory
+import piperkt.common.ddd.Factory
 import piperkt.common.id.ChannelId
 import piperkt.services.servers.domain.Channel
 import piperkt.services.servers.domain.ChannelType
@@ -12,7 +12,7 @@ object ChannelFactory : Factory<Channel> {
         name: String,
         description: String,
         id: String = ChannelId().value,
-        messages: MutableList<Message> = mutableListOf()
+        messages: MutableList<Message> = mutableListOf(),
     ): Channel {
         return Channel(
             name = name,
@@ -26,7 +26,7 @@ object ChannelFactory : Factory<Channel> {
     private fun createMultimediaChannel(
         name: String,
         description: String,
-        id: String = ChannelId().value
+        id: String = ChannelId().value,
     ): Channel {
         return Channel(
             name = name,
@@ -41,7 +41,7 @@ object ChannelFactory : Factory<Channel> {
         description: String,
         type: String,
         id: String = ChannelId().value,
-        messages: MutableList<Message> = mutableListOf()
+        messages: MutableList<Message> = mutableListOf(),
     ): Channel {
         if (ChannelType.valueOf(type) == ChannelType.TEXT) {
             return createMessageChannel(name, description, id, messages)
