+++ servers-service/src/main/kotlin/piperkt/services/servers/domain/factory/MessageFactory.kt
@@ -1,7 +1,7 @@
 package piperkt.services.servers.domain.factory
 
 import java.time.Instant
-import piperkt.common.Factory
+import piperkt.common.ddd.Factory
 import piperkt.common.id.MessageId
 import piperkt.services.servers.domain.Message
 
@@ -11,7 +11,7 @@ object MessageFactory : Factory<Message> {
         content: String,
         sender: String,
         timeStamp: String = Instant.now().toString(),
-        id: String = MessageId().value
+        id: String = MessageId().value,
     ) =
         Message(
             sender = sender,
