+++ users-service/src/test/kotlin/mocks/publishers/MockedUserEventPublisher.kt
@@ -1,6 +1,6 @@
 package mocks.publishers
 
-import piperkt.common.MockedEventPublisher
+import piperkt.common.mocks.MockedEventPublisher
 import piperkt.services.users.domain.user.UserEvent
 import piperkt.services.users.domain.user.UserEventPublisher
